Make Reusable compoents in NMoq Pages, Added NMoq Arabic strings
