--- QatarMuseum/ReusableComponents/TravelCell/TravelCell.xib
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="14109" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14088"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" reuseIdentifier="travelCellId" id="JoX-NO-NG7" customClass="TravelCollectionViewCell" customModule="QatarMuseums" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="318" height="178"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <view key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" insetsLayoutMarginsFromSafeArea="NO">
-                <rect key="frame" x="0.0" y="0.0" width="318" height="178"/>
-                <autoresizingMask key="autoresizingMask"/>
-                <subviews>
-                    <imageView userInteractionEnabled="NO" contentMode="scaleAspectFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="default_imageX2" translatesAutoresizingMaskIntoConstraints="NO" id="HKz-GA-KMO">
-                        <rect key="frame" x="0.0" y="0.0" width="318" height="178"/>
-                    </imageView>
-                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="jtH-cc-TGl">
-                        <rect key="frame" x="12" y="161" width="294" height="0.0"/>
-                        <fontDescription key="fontDescription" type="system" weight="medium" pointSize="12"/>
-                        <color key="textColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <nil key="highlightedColor"/>
-                    </label>
-                </subviews>
-            </view>
-            <constraints>
-                <constraint firstItem="HKz-GA-KMO" firstAttribute="leading" secondItem="JoX-NO-NG7" secondAttribute="leading" id="Ake-dg-FBH"/>
-                <constraint firstItem="HKz-GA-KMO" firstAttribute="top" secondItem="JoX-NO-NG7" secondAttribute="top" id="Fig-gC-qhA"/>
-                <constraint firstItem="jtH-cc-TGl" firstAttribute="leading" secondItem="JoX-NO-NG7" secondAttribute="leading" constant="12" id="Lb6-px-xpI"/>
-                <constraint firstAttribute="bottom" secondItem="jtH-cc-TGl" secondAttribute="bottom" constant="17" id="Pxc-WU-o4V"/>
-                <constraint firstAttribute="trailing" secondItem="HKz-GA-KMO" secondAttribute="trailing" id="Z6R-6g-7Op"/>
-                <constraint firstAttribute="bottom" secondItem="HKz-GA-KMO" secondAttribute="bottom" id="bBP-0Q-uy4"/>
-                <constraint firstAttribute="trailing" secondItem="jtH-cc-TGl" secondAttribute="trailing" constant="12" id="tcX-Up-2rS"/>
-            </constraints>
-            <viewLayoutGuide key="safeArea" id="Hmw-mo-s8T"/>
-            <connections>
-                <outlet property="imageView" destination="HKz-GA-KMO" id="uJO-Xc-AP3"/>
-                <outlet property="titleLabel" destination="jtH-cc-TGl" id="tdq-aq-7xF"/>
-            </connections>
-            <point key="canvasLocation" x="-72" y="-70"/>
-        </collectionViewCell>
-    </objects>
-    <resources>
-        <image name="default_imageX2" width="235" height="120"/>
-    </resources>
-</document>
