+++ QatarMuseum/ar.lproj/Localizable.strings
@@ -103,7 +103,7 @@
 "INSTITUTION_4" = "تعليم متحف الفن الإسلامي";
 "INSTITUTION_5" = "تعليم المتحف";
 "INSTITUTION_6" = "المتحف الوطني";
-"INSTITUTION_7" = "3-2-1 QOSM Education";
+"INSTITUTION_7" = "التعليم بمتحف قطر الأولمبي والرياضي";
 "INSTITUTION_8" = "تقويم برامج الأسرة والمدرسة";
 "AGE_GROUP_1" = "المعلمين";
 "AGE_GROUP_2" = "جميع الفئات العمرية";
@@ -129,7 +129,7 @@
 "PGM_TYPE_8" = "ورشة العمل";
 
 "KNOWLEDGE_ACTION_TITLE" = "المعرفة بين يديك";
-"EDUCATION_DESCRIPTION" = "Our numerous programmes are designed for families, university students and adults \n\n Find out what's on offer and register ahead to secure your place.";
+"EDUCATION_DESCRIPTION" = "تم تصميم برامجنا العديدة والمتنوعة لتتناسب مع العائلات وطلاب الجامعات والكبار. تعرَّف على ما يقدمه المتحف من برامج تعليمية وسارع بالتسجيل للاستفادة من هذه البرامج. ";
 "EDUCATION_TEXT" = "تُتيح جميع برامج التعليم لدينا فرص تفاعلية. نحن نأمل بأن ترسخ هذه البرامج في ذاكرة المشاركين وتؤدي إلى تطوير مهاراتهم وقدراتهم الإبداعية.\nأما بالنسبة لمعلمي المدارس، فقد أعددنا ورش عمل ومنتديات ودورات تدريبية حسب الطلب والتي تتناسب مع إحتياجاتهم. نحن نركز أيضا على العمل مع الأطفال لتشجيعهم على إستكشاف العالم من حولهم، والانخراط والمشاركة فيه والتعبير عن أنفسهم من خلال الأنشطة الإبداعية.\nوتتماشى جميع أنشطتنا مع المعايير المهنية للمجلس الأعلى للتعليم الخاصة بالمعلمين وبمعايير المناهج الوطنية.";
 "TOUR_GUIDES" = "المرشد";
 "TOUR_GUIDE_TEXT" = "اغمر نفسك في تفاصيل القطع التي اختيرت خصيصا لك من قبل خبرائنا";
@@ -148,13 +148,13 @@
 "SELF_GUIDED_TEXT1" = "ربما تريد أن تجرب جولاتنا كما يحلو لك";
 "SELF_GUIDED_TEXT2" = "اغمر نفسك في تفاصيل القطعة التي اختيرت خصيصا لك من قبل خبرائنا.";
 "MIA_TOUR_GUIDE_DESC1" = "أهلا بك في متاحف قطر. يرجى اختيار المتحف لإستكشاف القطع المعروضة";
-"MIA_TOUR_GUIDE_DESC2" = "Scan the QR codes available on the galleries for more information.";
-"SCIENCE_TOUR_TEXT" = "In this tour, you will see and learn about the achievements and contributions of Muslim scientists through history. The eight objects have an obvious scientific purpose or show how science has influenced artistic methods.";
+"MIA_TOUR_GUIDE_DESC2" = "امسح رموز QR المتاحة في صالات العرض لمزيد من المعلومات.";
+"SCIENCE_TOUR_TEXT" = "سيتسنى لك خلال هذه الجولة، رؤية واستكشاف إنجازات ومساهمات العلماء المسلمين عبر التاريخ. للقطع الثمانية غرض علمي واضح أو أنها تظهر كيفية تأثير العلوم على الأساليب والأشكال الفنية.";
 "SCIENCE_TOUR_TITLE" = "جولة العلوم";
 "YET_NO_NOTIFICATION_MESSAGE" = "لا توجد إشعارات بعد.";
 "SOMETHING_WENT_WRONG_TITLE" = "حدث خطأ ما";
 "LOCATION_ERROR_MESSAGE" = "للأسف لقد تعذر فتح خرائط جوجل";
-"ANYSTRING" = "All";
+"ANYSTRING" = "الكل";
 "LEVEL_STRING" = "المستوى";
 "MIA_HIGHLIGHTS_TOUR" = "جولة مختارات متحف الفن الإسلامي";
 "CULTURE_DISCOUNT_LABEL" = "- خصم بنسبة 15٪ في QM CAFÉ'S على جميع المنتجات \n\n- خصم بنسبة 10٪ على العناصر في جميع متاجر هدايا QM (بدون حد أدنى من الشراء) \n\n - خصم بنسبة 10٪ في مطعم IDAM في وقت الغداء \n\n- استلام النشرة الإخبارية الشهرية للبقاء على إطلاع على عروض متاحف قطر والشركاء \n\n- الحصول على وصول أولي إلى مناقشات وحلقات الأعضاء فقط \n\n- الحصول على دعوات حصرية إلى مركز اتصال متاحف قطر ذو المواصفات العالمية من 8 صباحًا إلى 8 مساءً يوميًا";
@@ -220,26 +220,26 @@
 "TAP_TO_FLIP" = "اضغط لعرض الجهة الأخرى";
 "FLOOR" = "المستوى";
 "GALLERY" = "المعرض";
-"TOUR_TITLE" = "TOURS";
-"TOURS" = "Tours";
-"TRAVEL_ARRANGEMENTS" = "Travel";
-"PANEL_DISCUSSION" = "Activities";
+"TOUR_TITLE" = "الجولات";
+"TOURS" = "الجولات";
+"TRAVEL_ARRANGEMENTS" = "السياحة";
+"PANEL_DISCUSSION" = "الأنشطة";
 "EVENT_DATE" = "EVENT DATE";
-"DOWNLOAD_TEXT" = "Download Brochure for More Information";
-"NMoQ_TOUR_DESC" = "This tour has been desgined for introducing you to the exquisite art & culture of Qatar. \n\nThe tour for the day has been divided into two part, one for the morning and other for the evening. Please find the information below and register so that we can make required arrangements for you.";
-"DATE" = "DATE";
-"VENUE" = "VENUE";
-"SPECIAL_OFFER" = "SPECIAL OFFER";
-"CLAIM_OFFER" = "Claim Offer";
-"NUMB_OF_RESERVATIONS" = "You have booked ";
-"SPACES" = " spaces";
-"SPACE" = " space";
-"TOUR_SEAT_AVAILABILITY_STRING1" = "Hurry up only ";
-"TOUR_SEAT_AVAILABILITY_STRING2" = " spaces available.";
-"TOUR_SEAT_AVAILABILITY_STRING3" = " space available.";
-"NO_SEAT_AVAILABLE" = "Fully booked";
-"BOOK_TOUR_STRING" = "Book Tour";
-"CANCEL_BOOKING_STRING" = "Cancel Booking";
-"UNREGISTER_CONFIRMATION_MESSAGE" = "Are you sure you want to cancel this tour?";
-"REGISTER_GREETING_MESSAGE" = "Thank you for booking this tour. You will receive an email with further information.\n\nAny questions, please contact \nbooking@qm.org.qa \n or \n+974 4452 5555";
+"DOWNLOAD_TEXT" = "تحميل الكتيب لمزيد من المعلومات";
+"NMoQ_TOUR_DESC" = "ستطلعك هذه الجولة على سحر الفن والثقافة في قطر. تم تقسيم الجولة إلى جزئين، أحدهما في الصباح والآخر مساءً. يرجى الاطلاع على المعلومات أدناه والتسجيل حتى نتمكن من اتخاذ الترتيبات اللازمة لك.";
+"DATE" = "التاريخ";
+"VENUE" = "المكان / الموقع";
+"SPECIAL_OFFER" = "عرض خاص";
+"CLAIM_OFFER" = "طلب العرض";
+"NUMB_OF_RESERVATIONS" = "قمت بحجز ";
+"SPACES" = " أماكن";
+"SPACE" = " مكان";
+"TOUR_SEAT_AVAILABILITY_STRING1" = "سارع بالحجز .... ";
+"TOUR_SEAT_AVAILABILITY_STRING2" = " الأماكن المتاحة / أماكن متاحة";
+"TOUR_SEAT_AVAILABILITY_STRING3" = " مكان متاح";
+"NO_SEAT_AVAILABLE" = "محجوزة بالكامل ";
+"BOOK_TOUR_STRING" = "احجز جولة";
+"CANCEL_BOOKING_STRING" = "إلغاء الحجز";
+"UNREGISTER_CONFIRMATION_MESSAGE" = "هل أنت متأكد من أنك تريد إلغاء هذه الجولة؟";
+"REGISTER_GREETING_MESSAGE" = "شكرا لك على حجز هذه الجولة. سوف تتلقى رسالة بريد إلكتروني مع مزيد من المعلومات. لأي أسئلة ، يرجى التواصل على booking@qm.org.qa أو 5554 4452 974+";
